functions/requirements.txt
# Firebase Functions dependencies
firebase-functions>=0.4.0
firebase-admin>=6.4.0

# Google Cloud Vision API
google-cloud-vision>=3.5.0

# Image processing
Pillow>=10.2.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
